Improve request code
The content-type and content-length alone indicate whether a
body is present for request/4.

The host header can now be overriden.

A number of headers like transfer-encoding are automatically
deleted from the list of headers we receive. SPDY also deletes
connection, keep-alive and proxy-connection.

SPDY now sends the port in the :host header.