Add Gun Websocket handlers and protocol support
This functionality can be used to implement custom protocols
on top of Websocket, but may also be used to decode frame
contents on the fly if necessary.

The default_protocol option defines what module should be
used when no protocol was selected.

The protocols option is a list of key/value pairs used to
select the handler depending on the protocol that the server
accepted.

The feature is currently experimental.