Initial support for raw send/recv operations
Gun can now be used to send or receive arbitrary data in the
following scenarios:

* Directly after connecting to a server (this is not terribly
  useful but it works nevertheless due to the Gun architecture)

* After connecting through one or more Socks and/or HTTP proxies.
  This allows using Gun's proxy capabilities to access servers
  located beyond firewalls.

* After performing an HTTP/1.1 Upgrade. This allows using Gun
  to implement custom protocols that require upgrading from
  an HTTP/1.1 connection.

As there is still no support for HTTP/2 CONNECT for the time
being, there are no relevant streams attached to those use
cases and therefore the raw protocol currently expects users
to use 'undefined' as the StreamRef value. This is not a
final decision and will most likely produce a Dialyzer
warning at this time.