Default keepalive to infinity for all protocols
For HTTP/2 some servers are not fond of receiving pings,
especially if there's no stream at the time. (particularly
gRPC servers). This is an odd decision for sure. Rather
than trying to accomodate for their logic the ping is no
longer sent by default.

For Websocket sending a ping is a new feature in Gun 2.0
so I'm just being conservative there.