Fail fast while closing if reconnect is off
If a request/headers/connect/ws_upgrade is created when a connection is in state
'closing', such as after receiving an HTTP/2 GOAWAY frame or an HTTP/1.1
"Connection: close" header, an error message is sent back to the caller
immediately, if reconnect is off (that is if the option retry is set to 0).

This allows an application to retry the request on another connection without
waiting for all streams on the current connection to complete.