Add keepalive_tolerance http2 option
keepalive_tolerance: The number of unacknowledged pings that can be tolerated
before the connection is forcefully closed.

When a keepalive ping is sent to the peer, a counter is incremented and if this
counter exceeds the tolerance limit, the connection is forcefully closed. The
counter is decremented whenever a ping ack is received from the peer.

By default, the mechanism for closing the connection based on ping and ping ack
is disabled.